Rotation development
The next functions were built:
Continual Rotation - Sets all the legs to begin rotating and pushing the bot forward.
Stop Rotation - As the name goes, reduces the velocity down to 0. Uses SetLimit function.
Set Limit - takes the current angle of each leg and sets the angle to be at most (current +10), at least (current -2). This is to have enough of a window for the engine to not send the legs flying and instead lock them in their current position.

Additional feature - Leg groups are relevant now and decide which leg starts inverted.